requirements.txt
certifi==2020.6.20
Pillow==8.4.0
pip==21.2.4
pygame==2.1.0
setuptools==58.0.4
wheel==0.37.0
